fix: correção do tipo de double para float

+++ src/main/java/com/pht/rinha_backend_2026/constants/Constants.java
@@ -1,7 +1,7 @@
 package com.pht.rinha_backend_2026.constants;
 
 public class Constants {
-    public  double max_amount = 10000;
+    public  float max_amount = 10000f;
     public  int max_installments = 12;
     public  int amount_vs_avg_ratio = 10;
     public  int max_minutes = 1440;
